+++ src/connection/base.js
@@ -20,7 +20,14 @@ export default class ConnectionBase {
   _pc: window.RTCPeerConnection.prototype;
   _callbacks: Object;
 
-  constructor(signalingUrl: string, role: string, channelId: string, metadata: string, options: ConnectionOptions, debug: boolean) {
+  constructor(
+    signalingUrl: string,
+    role: string,
+    channelId: string,
+    metadata: string,
+    options: ConnectionOptions,
+    debug: boolean
+  ) {
     this.role = role;
     this.channelId = channelId;
     this.metadata = metadata;
@@ -223,7 +230,11 @@ export default class ConnectionBase {
 
   _createAnswer(message: Object) {
     // simulcast の場合
-    if (this.options.simulcast && this.role === 'upstream' && message.encodings) {
+    if (
+      this.options.simulcast &&
+      (this.role === 'upstream' || this.role === 'sendrecv' || this.role === 'sendonly') &&
+      message.encodings
+    ) {
       const transceiver = this._pc.getTransceivers().find(t => {
         if (t.mid && 0 <= t.mid.indexOf('video') && t.currentDirection == null) {
           return t;